gcloudex v0.4.3 GCloudex.Auth

Provides retrieval of authentication tokens for several Google Cloud Platform services’ scopes.

Functions

get_token_storage(type)

Specs

get_token_storage(atom) :: binary

Retrieves an authentication token for the Google Cloud Storage service.